LayoutTests/editing/pasteboard/copying-html-strips-javascript-url-and-event-handler-expected.txt

This tests copying and pasting a markup with an event handler and a javascript URL. WebKit should strip them away upon copy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS html.includes("hello") is true
PASS html.includes("world") is true
PASS html.includes("dangerousCode") is false
PASS successfullyParsed is true

TEST COMPLETE
